consolidate pages and header links
